+++ src/frontend/dom/gui-components.js
@@ -1,5 +1,5 @@
-import { html } from "./utils";
-import Cmp from "./accessors";
+import { html } from "../utils";
+import { Cmp } from "./accessors";
 
 /**
  * @typedef GitMenuFunctions
@@ -35,7 +35,6 @@ class FileMenu {
   }
 }
 
-/** Git menu instantiator from Edit menu */
 class GitMenu {
   constructor() {
     this.menu = document.querySelectorAll(`div.${Cmp.MENU_ITEM}`)[2];
@@ -142,63 +141,53 @@ class GitMenu {
   }
 }
 
-class Alert {
-  static CLOSE_BUTTON_SVG =
-    "data:image/svg+xml;base64, \
-  PHN2ZyBpZD0iTGF5ZXJfMSIgZGF0YS1uYW1lPSJMYXllciAxIiB4bWxucz0iaHR0cDovL3d \
-  3dy53My5vcmcvMjAwMC9zdmciIHZpZXdCb3g9IjAgMCA3LjQ4IDcuNDgiPjxkZWZzPjxzdH \
-  lsZT4uY2xzLTF7ZmlsbDpub25lO3N0cm9rZTojZmZmO3N0cm9rZS1saW5lY2FwOnJvdW5kO \
-  3N0cm9rZS1saW5lam9pbjpyb3VuZDtzdHJva2Utd2lkdGg6MnB4O308L3N0eWxlPjwvZGVm \
-  cz48dGl0bGU+aWNvbi0tYWRkPC90aXRsZT48bGluZSBjbGFzcz0iY2xzLTEiIHgxPSIzLjc \
-  0IiB5MT0iNi40OCIgeDI9IjMuNzQiIHkyPSIxIi8+PGxpbmUgY2xhc3M9ImNscy0xIiB4MT \
-  0iMSIgeTE9IjMuNzQiIHgyPSI2LjQ4IiB5Mj0iMy43NCIvPjwvc3ZnPg==";
+const CLOSE_BUTTON_SVG =
+  "data:image/svg+xml;base64, \
+PHN2ZyBpZD0iTGF5ZXJfMSIgZGF0YS1uYW1lPSJMYXllciAxIiB4bWxucz0iaHR0cDovL3d \
+3dy53My5vcmcvMjAwMC9zdmciIHZpZXdCb3g9IjAgMCA3LjQ4IDcuNDgiPjxkZWZzPjxzdH \
+lsZT4uY2xzLTF7ZmlsbDpub25lO3N0cm9rZTojZmZmO3N0cm9rZS1saW5lY2FwOnJvdW5kO \
+3N0cm9rZS1saW5lam9pbjpyb3VuZDtzdHJva2Utd2lkdGg6MnB4O308L3N0eWxlPjwvZGVm \
+cz48dGl0bGU+aWNvbi0tYWRkPC90aXRsZT48bGluZSBjbGFzcz0iY2xzLTEiIHgxPSIzLjc \
+0IiB5MT0iNi40OCIgeDI9IjMuNzQiIHkyPSIxIi8+PGxpbmUgY2xhc3M9ImNscy0xIiB4MT \
+0iMSIgeTE9IjMuNzQiIHgyPSI2LjQ4IiB5Mj0iMy43NCIvPjwvc3ZnPg==";
 
-  /** @param {{message: string; duration: number}} */
-  constructor({ message, duration }) {
-    this.message = message;
-    this.duration = duration;
-  }
-
-  display() {
-    document.querySelector(`.${Cmp.ALERT_CONTAINER}`).innerHTML = html`<div
-      class="${Cmp.ALERT_DIALOG} ${Cmp.ALERT_SUCCESS} ${Cmp.BOX}"
-      style="justify-content: space-between"
-    >
-      <div class="${Cmp.ALERT_MESSAGE}">${this.message}</div>
-      <div class="${Cmp.ALERT_BUTTONS}">
-        <div class="${Cmp.ALERT_CLOSE_CONTAINER} ${Cmp.BOX}">
-          <div
-            aria-label="Close"
-            class="${Cmp.CLOSE_BUTTON} ${Cmp.CLOSE_BUTTON_LARGE}"
-            role="button"
-            tabindex="0"
-          >
-            <img
-              class="${Cmp.CLOSE_ICON} undefined"
-              src="${Alert.CLOSE_BUTTON_SVG}"
-            />
-          </div>
+/** Displays a custom scratch-gui alert
+ * @param {{message: string; duration: number}} */
+export function scratchAlert({ message, duration }) {
+  document.querySelector(`.${Cmp.ALERT_CONTAINER}`).innerHTML = html`<div
+    class="${Cmp.ALERT_DIALOG} ${Cmp.ALERT_SUCCESS} ${Cmp.BOX}"
+    style="justify-content: space-between"
+  >
+    <div class="${Cmp.ALERT_MESSAGE}">${message}</div>
+    <div class="${Cmp.ALERT_BUTTONS}">
+      <div class="${Cmp.ALERT_CLOSE_CONTAINER} ${Cmp.BOX}">
+        <div
+          aria-label="Close"
+          class="${Cmp.CLOSE_BUTTON} ${Cmp.CLOSE_BUTTON_LARGE}"
+          role="button"
+          tabindex="0"
+        >
+          <img
+          <!-- todo: why does this have an undefined class -->
+          class="${Cmp.CLOSE_ICON} undefined" src="${Alert.CLOSE_BUTTON_SVG}" />
         </div>
       </div>
-    </div>`;
+    </div>
+  </div>`;
 
-    if (document.querySelector("body").getAttribute("theme") === "dark") {
-      document.querySelector(`.${Cmp.CLOSE_BUTTON}`).style.backgroundColor =
-        "rgba(0, 0, 0, 0.255)";
-    }
-    document.querySelector(`.${Cmp.CLOSE_BUTTON}`).onclick = this.remove;
-    setTimeout(this.remove, this.duration);
+  if (document.querySelector("body").getAttribute("theme") === "dark") {
+    document.querySelector(`.${Cmp.CLOSE_BUTTON}`).style.backgroundColor =
+      "rgba(0, 0, 0, 0.255)";
   }
-
-  remove() {
-    document.querySelector(`.${Cmp.ALERT_CONTAINER}`).innerHTML = "";
-  }
+  document.querySelector(`.${Cmp.CLOSE_BUTTON}`).onclick = this.remove;
+  setTimeout(
+    () => (document.querySelector(`.${Cmp.ALERT_CONTAINER}`).innerHTML = ""),
+    this.duration
+  );
 }
 
-/** @param {{message: string; duration: number}} */
-export function scratchAlert({ message, duration }) {
-  new Alert({ message, duration }).display();
-}
-
+/** Manages functions with the file menu */
 export const fileMenu = new FileMenu();
+
+/** Manages the intialization of the Git menu */
 export const gitMenu = new GitMenu();
