+++ src/frontend/dom/index.js
@@ -1,0 +1,4 @@
+import { Cmp } from "./accessors";
+import { gitMenu, fileMenu } from "./gui-components";
+
+export { Cmp, gitMenu, fileMenu };
